Editor / Associated resource / Don't allow removal of siblings relation.
When 2 records A and B points to the same parent C. The associated panel will list A and B as siblings.
Therefore when editing A or B, user has no privilege to remove the relation in the siblings to the parent.
Also fix list of related records when a record has both siblings and aggregatedResource (only the last category was listed).