Reorder the elements for the `series` and `dataset` pages. (#6252)
* Reorder the elements for the `series` and `dataset` pages.

All layout related elements (the columns) are moved to the `series` and `dataset` template. All other templates have elements without layout info, so they can be moved freely without breaking the layout and columns.

Extra: Added 2 general classes for 'css' columns: `col-2` and `col-3`. These are used for the technical information, so when you place this information in a `col-md-12` you can add the class `col-3` to display all the list elements in 3 columns

* Record view / Avoid unneeded lines and restore 3 columns grouped for technical info.

* Add extra padding at the bottom of some sections

Co-authored-by: Francois Prunayre <fx.prunayre@gmail.com>