Externalise the configuration of targetUri property
The targetUri property of proxyURITemplateProxyServlet can be now defined externally to web.xml
in case the value stored there is empty.
The lookup order in that case would be:
1. Environment variable
2. System Property
3. Property in config.properties

The mechanism is similar to the one used to set the data directory, looking first for properties
prefixed by the web application name and the servlet name in web.xml and if nothing is found using
geonetwork as application name.