Change Diff to use org.bitbucket.cowwoc:diff-match-patch:1.2 dependency from maven central
This is a fork of google code diff-match-patch:diff-match-patch:current adapted to java naming convetions and available from maven central. It is by far the most popular fork of this algorithm.

By adopting this dependency we avoid the need to use https://maven.repository.redhat.com/ga/ improving build times.