DOI / Configure DOI URL pattern (#6347)
* DOI / Configure DOI URL pattern

When creating a DOI, the DOI id was created using `https://doi.org/{{prefix}}/{{uuid}}`

@pvgenuchten tested to customize the DataCite prefix in https://github.com/geonetwork/core-geonetwork/pull/6148
but as discussed it can be more flexible to add a new settings to customize the DOI with a pattern.

Default is `{{uuid}}` but the DOI structure can be customized with database id and/or record group eg. `example-{{groupOwner}}-{{id}}` or simply `{{id}}`. Some users complain about the length of current DOI - so instead of relying on https://shortdoi.org/ user could use the internal db id if it is unique in the prefix scope ...

An API operation to check the DOI created is added.

* DOI / Configure DOI URL pattern / Use default utils.