Modal popup window for Logo Picker (#5933)
* Modal popup window for Logo Picker

Related to https://github.com/geonetwork/core-geonetwork/pull/5855

The old logo selection had a dropdown with small icons. This proves to be difficult with large numbers of logos. A modal popup with the ability to change icon sizes solves this.

Other small UI changes:
- position/order of selected logo and buttons
- max width for selected logo

* Remove the preselected logo because it only works for the harvester

* If there's no logo selected, don't display a default logo on the select `option`/`button`. This is confusing, because this default icon is not linked to the object in qursion.

If there's logo, also display it on the selection `option`/`button`

* i18n / Fix JSON

Co-authored-by: Jose García <josegar74@gmail.com>
Co-authored-by: François Prunayre <fx.prunayre@gmail.com>