Citation / Refactor to be able to use it in other XSLT
Follow up of https://github.com/titellus/core-geonetwork/pull/41

The main goal here is to be able to add citation in other pages eg.
landing page. The citation formatter is reworked so that it can be
imported and used (like JSON-LD formatter).

eg.

```xsl

  <xsl:include href="../../formatter/citation/base.xsl"/>
  <xsl:include href="../../../iso19115-3.2018/formatter/citation/common.xsl"/>

  <xsl:variable name="citationInfo">
    <xsl:call-template name="get-iso19139-citation">
      <xsl:with-param name="metadata" select="$metadata"/>
      <xsl:with-param name="language" select="$language"/>
    </xsl:call-template>
  </xsl:variable>

  <xsl:apply-templates mode="citation" select="$citationInfo"/>
```

Also add the possibility to have an additional citation in the HTML
format. The additional citation is stored in the DOI online resource
description and indicate extra related work(s) to cite.