API / XPath editing / Delete and replace_all fix (#6506)
* API / XPath editing / Delete non existent node

Using XPath API to update a record was creating empty node when deleting non existent node (usually when xpath contains filter with no match). In deletion mode, we should never try to create elements corresponding to the requested XPath if there is no match.

* API / XPath editing / gn_replace_all mode may keep empty elements.

Issue introduced in
https://github.com/geonetwork/core-geonetwork/pull/6423.

Symptoms : edit a record with a topic category, switch to xml view
Empty `<gmd:topicCategory/>` elements are in the record.

When `gn_replace_all` is used the `gn_delete` steps was not take into
account.