[BP] Enable server hostname verification on this SSL/TLS connection. (#6403)
* Enable server hostname verification on this SSL/TLS connection.

To establish a SSL/TLS connection not vulnerable to man-in-the-middle attacks,
it’s essential to make sure the server presents the right certificate.

The certificate’s hostname-specific data should match the server hostname.

* Verify the mail server identity

If using TLS or SSL verify the mail server identity if no checked the
ignore certificate errors setting.