Search / Aggregation / Add decorator
Currently only home page aggregations INSPIRE themes, ISO topics and resource type are illustrated with an icon.
This PR makes it more generic so that we can configure the way we want to decorate an aggregation in the home page or in other pages.

A decorator can be used to add an icon or image next to an aggregation. The decorator is configured in the `meta` properties of the facet:

```js
"resourceType": {
  "terms": {
    "field": "resourceType",
    "size": 10
  },
  "meta": {
    "decorator": {
      "type": "icon",
      "prefix": "fa fa-2x pull-left gn-icon-"
    }
}
```

Decorator types are:

* Fixed icon. eg. adding a tag icon to all values

```js
decorator: {
  type: 'icon',
  prefix: "fa fa-fw fa-tag "
}
```

* Icon defined in a CSS class name using the value. eg. used for resource types `gn-icon-dataset`

```js
"decorator": {
  "type": "icon",
  "prefix": "fa fa-fw gn-icon-"
}
```

* Icon defined in a css class name which is using only a portion of the value. eg. used for INSPIRE themes. The expression allows to extract the value

```js
"decorator": {
  "type": "icon",
  "prefix": "fa fa-fw gn-icon iti-",
  "expression": "http://inspire.ec.europa.eu/theme/(.*)"
}
```

* Icon defined with a map of values for the class name to use.

```js
"decorator": {
  "type": "icon",
  "prefix": "fa fa-fw ",
  "map": {
    "availableInViewService": "fa-globe",
    "availableInDownloadService": 'fa-download"
  }
}
```

Define the image to use for each values:

```js
"decorator": {
  "type": "img",
  "map": {
    "EEA": "https://upload.wikimedia.org/wikipedia/en/thumb/7/79/EEA_agency_logo.svg/220px-EEA_agency_logo.svg.png"
  }
}
```