Fix JS exception when user info is not loaded yet
Sometime the user information is not loaded yet when the loadSettings function is called, causing
a Javascript exception when checking for the user.isAdministratorOrMore(). This commit removes the
guard requesting build info and loggging from the API for all users with access to the settings page.
The server will return a 403 Forbidden for this calls showing the warning in the developers tools
console but the Javascript will not fail.