New metadata / Metadata identifier template improvements.
In admin console, add a link to the documentation.

Create a directive for the component in charge of building the identifier from the template and reuse it in horizontal and vertical form.

All templates use a non empty template so same rules can apply everywhere.

Add support for UUID in template in order to compute a random UUID in a template with a specific scheme.