API / Batch edit / Delete with a non matching XPath can remove unwanted elements.
When using gn_delete tag, always consider it as an XML mode to properly
check that we are in deletion mode in EditLib.

Fix wrong test and test various type of config can be self closing tag
or not <gn_delete/> or <gn_delete></gn_delete>.

Issue was noticed using the following config

{"field":"Distribution / Remove an online resource by URL","insertMode":"gn_delete","xpath":"/gmd:MD_Metadata/gmd:distributionInfo/*/gmd:transferOptions/*/gmd:onLine[starts-with(*/gmd:linkage/gmd:URL, \"http://www.eea.europa.eu/data-and-maps/data\") or starts-with(*/gmd:linkage/gmd:URL, \"https://cmshare.eea.europa.eu/\")]","value":"","isXpath":true}

which was removing the first gmd:onLine when none were matching.