Record view / Display bounding polygon if set
Most of the time only a bounding box is provided in metadata records but sometimes bounding polygons are also used.
Add the possibility to return the bounding polygon if available for a metadata record.
For the time being, this is only rendered in the record view. 

For search results, only the bounding box is rendered (as bounding polygon can increase a lot the search response size). If some user wants to highlight polygons in the small map, they can add the `shape` field in the `gnESFacet` include list.