Aggregation / Translate bucket value based on search field
User may configure a custom facet with some include/exclude config to
only collect statistics on some values eg.

```json
{
  "onlySomeGroup": {
    "terms": {
      "field": "groupOwner",
      "size": 80,
      "include": "101|102",
      "order": {
        "_key": "asc"
      }
    },
    "meta": {
      "orderByTranslation": true,
      "field": "groupOwner"
    }
  },
```

In this case, values 101 and 102 are translated client side based on
group labels. As the facet has a custom name (ie. `onlySomeGroup`)
group id are not translated. The translation has to be based on the
meta field name in this case.