Permalink / Set link URL as in sitemap.
Follow up of https://github.com/geonetwork/core-geonetwork/pull/5579.

The same URL is used in sitemap and on client side permalink popup.
Default permalink is a link pointing to the API (and not the client app
with the UUID after the # which makes the information not available
server side and no redirect easy to setup).

The permalink points to an HTML page containing the JSON-LD
representation needed for structured data indexing for SEO.

Also relate to discussion in https://github.com/geonetwork/core-geonetwork/issues/6553