Harvesting / Features / Check if index exist (#6492)
In most cases when app starts, the records and features index are created.

But if feature index is removed manually before indexing features, then the index schema will not be the one defined in config (it will be automatically created by Elasticsearch). Symptom of this is that aggregations on data preview are empty.

Add a check to init feature index if needed before harvesting a layer.