Aggregation / Add a refresh policy options.
Current general mechanism was to not refresh the last selected facets with some exception for nested facet and tree.
But when selecting facet A, then B, then A again - both A and B have been refreshed and some users are not really convinced by that approach.

Current approach also has an issue when reloading the page with a facet selected, users would like for some important facet to display all values and not only the selected one. eg. select resource type = dataset, reload the page, only dataset is visible.

This turn off the last facet is not refreshed rule and adds a refresh policy property so that user can defined on a per facet basis the way it behaves. For now, all facets refresh others but if refreshPolicy is set to displayAll, the initial values are displayed (with no count). So user can continue selecting values in those facets and do an OR query on all selected values.