Adds a profile to prebuilt the wro4j-cache
Add the wro4j-prebuild-cache maven profile for building the WRO4J cache in
build time and use it when Geonetwork starts.
If the data directory doesn't contain a wro4j-cache database or the git version
stored in the data directory doesn't match the GeoNetwork's git commit hash then
copy the prebuilt cache to the data directory.
If copied the cache and GN has custom CSS & JS settings in the DB then delete
all the CSS items from the cache to allow the LESS compiler take it on runtime.