Permalink improvement
Make more consistent all links to metadata record.

Follow up of
https://github.com/geonetwork/core-geonetwork/pull/5579
https://github.com/geonetwork/core-geonetwork/pull/6738

Basically we have 2 type of links for record:
* one for SEO in sitemap and permalink which return HTML (+ JSON-LD)
* one for the JS app (ie. catalog.search).

Those settings are moved to the metadata section.

In some environement user may use DOI and want the canonical URL to be the DOI. An option is added to the settings to have DOI take precedence.

In the API a new operation is added to retrieve the permalink (so it is consistent in XSL or JS app and easier to retrieve).

The JS app link is used in formatter which may provide links to the portal. Add utilities to retrieve it and add a new setting to be able to override it (in case a custom app is used as the main entry point for the catalogue content eg. sextant).

In notification email use the JS app link too in all places.