Add a setting to make the map accessible. The setting is off by default.
The accessible map caused some confusion to users because you first need the focus on the map before you can interact with it. Zooming in and out with the mouse doesn't work without the map having focus.

The new setting reverts the map to the old situation by default (direct interaction) and can be switeched on to have keyboard interaction when the map has the focus.