Map / Context / Style may or may not declare namespace.
In GeoServer, GetCapabilities may declare layer and style name with
namespace when querying the global capabilities but not when querying a
workspace capabilities (not sure if this has changed recently or if it a
configuration?).

Make style name search more efficient to match style with namespace or
not. Some past saved map have been saved with workspace URL with style
name with namespace and they now failed to load properly.