Thesaurus / Virtual all thesaurus modifying Keyword bean in cache
When using virtual all thesaurus, keyword can be added with a reference
to the virtual all thesaurus instead of the thesaurus of the keyword
(ie. no thesaurus date and title set to All keywords).

Also when using XLinks, the dispatching of keywords based on their
thesaurus of origin was building xlink URL with concept id as thesaurus
parameter.

Remove unecessary duplicated skipdescriptivekeyword parameter when
dispatching keyword from the all thesaurus.

For testing:
* enable virtual all thesaurus in settings and xlink
* try to add keywords (with xlink mode)
Keywords are not added properly.

Workaround for previous version:
* set thesaurus.cache.maxsize=0 in config.properties.