CSW / Transaction / Consistently apply update-fixed-info
`Update-fixed-info` was only applied on `Insert` (but not on `Update`).

This was causing an issue on multilingual records as on Insert all namespaces were located on the root element but on Update namespaces were declared when used.

On multilingual records, `lan` namespace was not declared when used on elements declaring `xsi:type` attribute and validation was failing on that. The dispatching of namespaces sounds more related to a JDOM issue not taking into account `xsi:type`.

In any case, it make more sense to be consistent on applying `update-fixed-info` like when using the default open API. The db settings configure if it has to be applied or not.