DOI / mEDRA support / Fix issue with DataCite publication.
When a namespace prefix is set for the `datacite:resource`
mEDRA manage to publish the doi and the https://doi.org... URL
is working fine but the dashboard history if the OP DOI Registration
Agency is not reporting green status for the 'result on Datacite'
column.

Removing the namespace prefix and using a default namespace only for the
`resource` element, so replacing
```
<eu:Metadata>
<datacite:resource xmlns:datacite="http://datacite.org/schema/kernel-4">
```
by
```
<eu:Metadata>
<resource xmlns="http://datacite.org/schema/kernel-4">
```
make it works even if exactly the same from an XML point of view.