Config / Add property file overlay.
We often have the need to adapt application properties.

Some of the properties are customized using env variables eg.
or using Java arguments eg.

In docker,  is also used to set variables in.

Add the possibility to apply overlay on top of default properties using property files stored in data directory:
* config/app.properties
* config/db.properties
* config/security.properties
so they can persist on update and only contains properties which needs a specific value.