Editing / dateStamp = changeDate on update (#7012)
* Editing / dateStamp = changeDate on update

On metadata update, `BaseMetadataManager` is calling `update-fixed-info` which set `dateStamp` from metadata table `changeDate` but the table is only updated after in `XmlSerializerDb`. The dateStamp is using the previous changeDate (usually a short time before last edit on an editing session with multiple actions). If user just open the editor and save and close directly, the dateStamp is set to the last editing date (which could be long time ago).

In the index, we expect to have `dateStamp` = `changeDate` but it was not the case.

* Update core/src/main/java/org/fao/geonet/kernel/datamanager/base/BaseMetadataManager.java

Co-authored-by: Jose García <josegar74@gmail.com>

---------

Co-authored-by: Jose García <josegar74@gmail.com>