Workflow / Remove magic numbers in label and avoid issue when using custom status.
The status workflow index field contains 2 numbers concatenated - status of the record + status of the draft.
If using custom status, the concatenation of the number can match other status number eg. 111 can be 1+11 or 11+1 so we can't always get a label from that value.
Use status name instead of ids and rely on db translation as much as
possible (there is still room for improvements when custom status are
used).

Co-authored-by: Joachim Nielandt <joachim.nielandt@vlaanderen.be>