Fix microservices proxy detection
Follow-up of https://github.com/geonetwork/core-geonetwork/pull/7094.

The targetUri property can be configured using multiple methods, not only as init parameter
in web.xml servlet's config but also as environment variable, system property or config.properties
entry. This change saves the targetUri value in a ServletContext attribute so it can be retrieved
from other servlets to check its value. Changed the value of <load-on-startup> servlet property
to init the attribute on server start.