Map / Layer manager / Add support for multilingual layer title.
Metadata may be multilingual and provide multilingual layer name in the
online source description field. GetCapabilities document do not provide
multilingual support, so this adds the possibility to translate layer
name in maps.

A multilingual description in a link:

```xml
<gmd:onLine>
  <gmd:CI_OnlineResource>
    <gmd:linkage>
      <gmd:URL>https://data.apps.fao.org/map/gsrv/gsrv1/geonetwork/ows?SERVICE=WMS&amp;</gmd:URL>
    </gmd:linkage>
    <gmd:protocol>
      <gmx:MimeFileType type="">OGC:WMS</gmx:MimeFileType>
    </gmd:protocol>
    <gmd:name xsi:type="gmd:PT_FreeText_PropertyType">
      <gco:CharacterString>geonetwork:basins_296</gco:CharacterString>
      <gmd:PT_FreeText>
        <gmd:textGroup>
          <gmd:LocalisedCharacterString locale="#EN">geonetwork:basins_296</gmd:LocalisedCharacterString>
        </gmd:textGroup>
        <gmd:textGroup>
          <gmd:LocalisedCharacterString locale="#FR" />
        </gmd:textGroup>
      </gmd:PT_FreeText>
    </gmd:name>
    <gmd:description xsi:type="gmd:PT_FreeText_PropertyType">
      <gco:CharacterString>Hydrological basins in Africa</gco:CharacterString>
      <gmd:PT_FreeText>
        <gmd:textGroup>
          <gmd:LocalisedCharacterString locale="#EN">Hydrological basins in Africa</gmd:LocalisedCharacterString>
        </gmd:textGroup>
        <gmd:textGroup>
          <gmd:LocalisedCharacterString locale="#FR">Bassins versants en Afrique</gmd:LocalisedCharacterString>
        </gmd:textGroup>
      </gmd:PT_FreeText>
    </gmd:description>
  </gmd:CI_OnlineResource>
</gmd:onLine>
```

The description is retrieved from the metadata on load from the metadata
UUID which is attached to the layer (and saved in the OWS context
extension of the layer) using the service URL and the layer name.

This also allows to override layer name when a group of layer (eg.
`a,b,c`) is used in the online source name.