Map / Background layer / Fix active layer on load.
Test:
* Open the map
* Reload page
* Active background is not enabled.

This was related to the fact that the search map is also affecting bgLayers.

Only the main map viewer is defining the bgLayers which is used for the base switcher directive.
In case of the context has no active background, use the first one (and log a warning).
Reset default map :
* may fail if no context defined. Add a fallback to the default one (and log a warning).
* tried to use the map layer config if no context defined as it was ignored (but not working - require more work or removal of those unused ? settings)