[BP] Send metadata mail notifications for public metadata when is re-approved (#7074)
When a metadata is published / un-published it can be configured a mail notification.

When the metadata workflow is enabled, published metadata can be edited and re-approved.
This change request expands the email notifications in this case, to send an email to let you know
that a published metadata has been re-approved, with a new version published.

Related to #6457

---------

Co-authored-by: Juan Luis Rodríguez <juanluisrp@gmail.com>