API / Import record / Overwrite mode improvement.
Currently `OVERWRITE` mode is not really overwriting the record
but remove it (including status, privileges, ...) and then replace it
with the new XML and import parameters.

Users expect that overwrite is only updating the XML of the metadata
record.

Change `OVERWRITE` mode to only update the XML and
add a `REMOVE_AND_REPLACE` mode to keep existing mechanism.