Remote INSPIRE ATOM Feed / code cleanup and optimisations for ATOM describe service (#7189)
Code optimisations to retrieve the atom feeds. Tested in 3.12.x and somehow retrieveDatasetUuidFromIdentifier 
causes that the call to findByMetadataId takes long time. In main doesn't seem so, maybe due to the 
JPA/Hibernate version.

The change does some cleanup of the code and avoids to retrieve the same information several times.