Standard / ISO19115-3 / Do not declare main language in other languages (#7188)
* Standard / ISO19115-3 / Do not declare main language in other languages

In ISO19139, the main language was also declared as a `locale` object in order to set the ID for the main language used for translations. See https://github.com/geonetwork/core-geonetwork/blob/main/schemas/iso19139/src/main/plugin/iso19139/update-fixed-info.xsl#L169-L186

In ISO19115-3, `defaultLocale` and `otherLocale` are both locale object so there is no need to declare main language in 2 places like in ISO19139. The ID is already set in the `defaultLocale` element.

This could happen when users migrate from ISO19139.

* Standard / ISO19115-3 / Multilingual / Do not fail if languages declared more than once