INSPIRE / Validation / Add API usage.
When using the JRC INSPIRE validator, users can follow their API usage and check if the quota of request is over or not.
The usage API return the number of validation per day. Usage seems to not be updated right after triggering validation, and the API reflect the usage after some time.