HA / Introducing scheduled refresh of harvesterjobs.
When using multiple nodes, one is in charge of harvesting tasks.
Use the following configuration for this node:
```
      HARVESTER_SCHEDULER_ENABLED: "true"
      HARVESTER_REFRESH_INTERVAL_MINUTES: 2
```
The node will then check every 2 minutes for any harvesting
configuration changes and updates its schedule.

This is a stopgap solution until we define a better messaging system to
deal with those cases.

Keep the default config if running only one node.

Co-authored-by: Joachim Nielandt <joachim.nielandt@vlaanderen.be>