Fix write after response commit on proxy response
When the Links table is empty an 500 error happens:
java.lang.IllegalStateException: Cannot call sendError() after the response has been committed

This is caused because the response.sendError(...) method is being called twice when
the Links table has no entries. This commits fix this error calling the method only
once.
In addition, it removes unnecessary info returned to the client.

Signed-off-by: Juan Luis Rodríguez <juanluisrp@gmail.com>