Add 404 error page
Github Pages needs a 404.html in the web root to show a custom 404 error page.
We could copy stable/404.html page to / but we would need to keep it updated
with styles and versions. Instead of that this commit creates a symbolic link
to /stable/404.html as suggested at https://github.com/orgs/community/discussions/23792.