[BP] INSPIRE Validator - exception handling improvements (#7519)
Refactor of the INSPIRE Validator backend code to improve the exception handling.

Previously some exceptions were not reported properly to the user. For example, if the upload of a file to validate returned an error, due to the validator not working properly (for example, returning an error 503), the user got an unclear error message.
Additionally this commit contains some fixes to Sonarlint warnings