[BP] Fix initialization of SpringDoc so that it does not use null servername and null version (#7575)
* Fix initialization of SpringDoc so that it does not use null servername and null version
Modified OpenApi object so that it is a static object that is shared with all context.  This helped fix initial api setting so that it was not set to null://null:80/geonetwork/srv/api
Also update SiteApi so that if the server properties are changed then it will update the OpenApi object without requiring an application restart.

* Update OpenAPI bean so that it calls the static method to ensure that no static values are changed directly in the bean.