Workflow / Editor board / Add search configuration to display draft, approved record or both (#7477)
* Workflow / Editor board / Display draft instead of approved records.

When using the workflow, if a published and approved record is edited, a
working copy is created. Users have difficulties to find the
corresponding working copies because the searches show the approved
records with a link to the copies but does not search directly for
working copies.

This change propose to have:
* the search app to search for records without draft and approved records
* the editor board app to search for records without draft and drafts.

In the editor board, sorting by last updates will properly list last
updated draft records and will facilitate editing works. The draft will
provide a link to the corresponding approved records (instead of
approved record linking to the working copy). It also makes more sense
to list draft here as user clicking edit button will edit the draft
record.

* Update web-ui/src/main/resources/catalog/components/search/resultsview/partials/viewtemplates/editor.html

Co-authored-by: joachimnielandt <joachim.nielandt@vlaanderen.be>

* Update web-ui/src/main/resources/catalog/components/search/resultsview/partials/viewtemplates/editor.html

Co-authored-by: joachimnielandt <joachim.nielandt@vlaanderen.be>

* Update web-ui/src/main/resources/catalog/components/utility/partials/draftvalidationwidget.html

Co-authored-by: joachimnielandt <joachim.nielandt@vlaanderen.be>

* Workflow / Editor board / Configure which type of record to promote in search - draft, approved or both.

* Workflow / Editor board / Configure which type of record to promote in search / Admin config.

* Workflow / Editor board / Prettier.

* Workflow / Editor board / Display approved status in badge.

* Workflow / Editor board / Configuration / Improve description.

---------

Co-authored-by: joachimnielandt <joachim.nielandt@vlaanderen.be>