Proxy / Make distinction between methods. (#101)
Some servers provides CORS header for some method eg. GET and not for other eg. HEAD.
In such case, if a failure occurs on a HEAD request, then the proxy is activated for that domain but is not needed for GET request. If user is anonymous, then proxy forbid access to the domain (if the domain is not in the link analysis table with default configuration) and simple action like `GetCapabilities` will fail.

Add the method to the domain as key to know if the proxy has to be used or not.

Can be tested with https://haleconnect.com/ows/services/org.789.b36c0053-db6b-4fe7-9402-4e5d8ac35e06_wms