Database / Default content / Remove categories.
The default categories provided in GeoNetwork since the early beginning are quite overlapping hierarchy level in ISO (eg. dataset, application, map) and in practice they are never used with the default values. It was making sense at the start to have a common grouping systems for the initial standards available in GeoNetwork like dublin core, FGDC, iso19115. Nowadays it is not recommended to use this default category list.

Here we propose to not populate the default database with any category (and we do not alter databases which may be using them).
This will also keep the UI simple. eg. If there is no category, no menu available in the editor (so users are not tempted to use the non recommended default values provided).

We should highlight that the main (only) benefit of using category mechanism is to group records without adding the information in the metadata document.