Improvements to the menu and footer on small screens.
The menu on small screens was complicated, didn't fit (padding and margins) and wasn't scrollable. These changes revert the menu to the original bootstrap menu. The styling also includes the portals that can be shown in the menu.

The footer on small screens had 'hidden' content, pages and social links were pushed off the screen when the footer was fixed. On smaal screens the footer isn't fixed anymore and displays the footer on more than one line.