Editor / Keywords / Add option to configure keyword encoding for a view.
More and more the default encoding for keywords is using `Anchor`
instead of `CharacterString`. It is even mandatory in a number of case
like INSPIRE validation.

Quite often, all editor views are configured using:

```xml
      <thesaurusList>
        <thesaurus key="external.theme.httpinspireeceuropaeutheme-theme"
                   transformations="to-iso19115-3.2018-keyword-with-anchor"/>
        <thesaurus key="external.theme.httpinspireeceuropaeumetadatacodelistSpatialScope-SpatialScope"
                   transformations="to-iso19115-3.2018-keyword-with-anchor"/>
      ...
```
to enable usage of `Anchor` for a number of thesaurus.

This change propose to be able to configure the default keyword encoding
for all thesaurus using:

```xml
      <thesaurusList defaultTransformation="to-iso19115-3.2018-keyword-with-anchor"/>
    </view>
```

If defined, it takes priority over the current encoding in the metadata
record.

It also hide the XLink encoding mode in the encoding selector if the
XLink resolver is not enabled.

Funded by Service Public de Wallonie