Editor / Add list mode (#7998)
* Editor / Add list mode

The list can be used to display a list of contact with only organisation name for example.
This can facilitate building simple form hiding the details of ISO elements. In this case,
use a contact directory to add new contact makes sense.

This mode can be also more efficient when record contains lots of contacts which in some case
can create large HTML form which can take time to build, download and render in the browser
(eg. when having more than 100 contacts in a record). If only `label` are used the form will be faster to render.

For large XML record, form may trigger error like `Form with too many keys [1001 > 1000]` when submitted.
The list mode allows to deal with such situation.

eg.

```xml
<action type="add"
        forceLabel="true"
        addDirective="data-gn-directory-entry-selector"
        or="pointOfContact"
        in="/gmd:MD_Metadata/gmd:identificationInfo/gmd:MD_DataIdentification">
  <directiveAttributes
          data-template-add-action="false"
          data-search-action="true"
          data-popup-action="true"
          data-template-type="contact"
          data-variables="gmd:role/gmd:CI_RoleCode/@codeListValue~{role}"/>
</action>

<list name="Contact point"
      xpath="/gmd:MD_Metadata/gmd:identificationInfo/*/gmd:pointOfContact[*/gmd:role/*/@codeListValue = 'pointOfContact']"
      del="."
      sortBy="gmd:organisationName/gco:CharacterString">
  <item href="*[1]/@uuid">
    <label xpath="*/gmd:organisationName/gco:CharacterString/text()"/>
  </item>
</list>

<list name="Originator"
      xpath="/gmd:MD_Metadata/gmd:identificationInfo/*/gmd:pointOfContact[*/gmd:role/*/@codeListValue = 'originator']"
      sortBy="gmd:organisationName/gco:CharacterString">
  <item>
    <field xpath="*/gmd:organisationName"/>
  </item>
</list>
```

Funded by Ifremer

* Editor / Add list mode / Display list as standard field in view mode 

This is needed when the custom editor is also used as a formatter.