[Backport 4.2.x] API / XSL / Use XSL output configuration (formatter, editor) (#8104)
* API / XSL / Use XSL output configuration (formatter, editor). (#7929)

* API / Formatter / Use XSL output configuration.

Formatter API was using an XMLOutputter formatter overriding all
configuration made in `xsl:output`. This was producing "invalid" HTML
and some workaround were used so far.

Now, return the output produced by saxon XMLIndenter or HTMLIndenter
directly.

* Editor / Fix formatting.

The editor is an HTML page so don't use an XML formatter for the output.
To format the XML in XML view, saxon is taking care of it using
serialize and default-indent-mode. Use an HTML `pre` element which is
not formatted by HTMLIndenter.

* Formatter / Fix conflicting output when citation is used in xsl-view.

* API / Formatter / iText for PDF expect xhtml document. If HTML, complains about not closed tags like links.

* Test / Fix formatter test. (#7993)

* Test / Fix formatter test.

Related to https://github.com/geonetwork/core-geonetwork/pull/7929.

* Prettier.

* Test / Fix formatter test / Citation common may be embedded in other formatter. Only set xsl:output in the citation entry point.

* API / XSL / Use XSL output configuration - formatter update for Java 8

---------

Co-authored-by: François Prunayre <fx.prunayre@gmail.com>